feat: integrated the 'login' functionality into the 'createUserProvider'
We kept the `login` function as a utility, but we also integrated into the `createUserProvider` to
make thigns easier. Due to this change we had to remove the `set` and `unset` methods from the
`Consumer`, instead we provide a `login`, `logout` and an `update` function where `e `update` is the
same as `set`, `logout` is the same as `unset` and `login` is the same as `login` utliity

BREAKING CHANGE: We kept the `login` function as a utility, but we also integrated into the
`createUserProvider` to make thigns easier. Due to this change we had to remove the